requirements.txt
cachetools==5.3.1
colorama==0.4.6
google-auth==2.20.0
google-auth-oauthlib==1.0.0
gspread==5.9.0
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyfiglet==0.8.post1
requests-oauthlib==1.3.1
rsa==4.9
